Clean up and make more hooks & context exportable, given integrators more flexibility when it comes to building on top of their own UI
