+++ copilot-widget/index.html
@@ -78,7 +78,7 @@
         headers: {
           // optional
           // Authorization: "Bearer BQAlIe479yo16aXf",
-          Cookie: `SESSIONID_MYM=pbljd8rnim9pt46d652olit663;`,
+          Cookie: `SESSIONID_MYM=6op56rogo023pqsas94tq6nbbl;`,
         },
         queryParams: {
           // optional
