+++ copilot-widget/lib/@components/Fallback.component.tsx
@@ -1,4 +1,4 @@
-import type { ComponentProps } from "../contexts/componentRegistery";
+import type { ComponentProps } from "@lib/types";
 
 type Props = ComponentProps<unknown>;
