+++ copilot-widget/lib/@components/index.ts
@@ -1,0 +1,4 @@
+export * from "./Fallback.component";
+export * from "./Handoff.component";
+export * from "./Loading.component";
+export * from "./Text.component";
