+++ copilot-widget/lib/components/index.ts
@@ -1,0 +1,12 @@
+export * from "./BotIcon";
+export * from "./BotMessage";
+export * from "./BotMessageWrapper";
+export * from "./Button";
+export * from "./ChatHeader";
+export * from "./ChatInputFooter";
+export * from "./Dialog";
+export * from "./InitialBotMessage";
+export * from "./UserAvatar";
+export * from "./UserMessage";
+export * from "./VoiceRecorder";
+export * from "./Vote";
