+++ copilot-widget/lib/contexts/WidgetState.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from "react";
-import useToggle from "../hooks/useToggle";
+import { useToggle } from "@lib/hooks";
 import { useConfigData } from "./ConfigData";
 import { createSafeContext } from "./createSafeContext";
 
@@ -11,10 +11,10 @@ const [useWidgetState, WidgetStateSafeProvider] =
 export default function WidgetState({ children }: { children: ReactNode }) {
   const { defaultOpen } = useConfigData();
   const data = useToggle(defaultOpen ?? false);
+
   return (
     <WidgetStateSafeProvider value={data}>{children}</WidgetStateSafeProvider>
   );
 }
 
-// eslint-disable-next-line react-refresh/only-export-components
 export { useWidgetState };
