+++ copilot-widget/lib/contexts/componentRegistry.ts
@@ -1,28 +1,12 @@
-import React from "react";
-import type { BotMessageType } from "./messageHandler";
-
-import { Fallback } from "@lib/@components/Fallback.component";
+import type { ComponentType, OptionsType } from "@lib/types";
+import { Fallback, Handoff, Loading, Text } from "@lib/@components";
 import { FormComponent } from "@lib/@components/Form.component";
-import { Loading } from "@lib/@components/Loading.component";
-import { Text } from "@lib/@components/Text.component";
-import { HandoffComponentRenderer } from "@lib/@components/Handoff.component";
-
-export type ComponentProps<TData> = BotMessageType<TData>;
-
-export type ComponentType = {
-  key: string;
-  component: React.ElementType;
-};
-
-type OptionsType = {
-  components?: ComponentType[];
-};
 
 /**
  * this a singleton  class helps me to easily control the components present/available in the widget.
- * it also manges the various types of components to be added along the way.
+ * it also manages the various types of components to be added along the way.
  */
-export class ComponentRegistery {
+export class ComponentRegistry {
   components: ComponentType[] = [
     {
       key: "TEXT",
@@ -42,7 +26,7 @@ export class ComponentRegistery {
     },
     {
       key: "HANDOFF",
-      component: HandoffComponentRenderer,
+      component: Handoff,
     },
   ] as const;
 
@@ -52,6 +36,7 @@ export class ComponentRegistery {
     if (components) {
       components.forEach((c) => this.register(c));
     }
+
     if (this.components.length === 0) {
       throw new Error("No components registered");
     } else if (!this.get("FALLBACK")) {
@@ -60,7 +45,7 @@ export class ComponentRegistery {
   }
 
   register(com: ComponentType) {
-    // Replace the key if already exists
+    // Replace the key if it already exists
     const index = this.components.findIndex((c) => c.key === com.key);
     if (index !== -1) {
       this.components[index] = com;
@@ -77,13 +62,14 @@ export class ComponentRegistery {
   }
 
   private getOrFallback(key?: string) {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     return key ? this.get(key) || this.get("FALLBACK")! : this.get("FALLBACK")!;
   }
+
   public getComponent(key: string, getFallback?: boolean) {
     if (getFallback) {
       return this.getOrFallback(key).component;
     }
+
     return this.get(key)?.component;
   }
 }
