+++ copilot-widget/lib/contexts/createSafeContext.ts
@@ -2,6 +2,7 @@ import { createContext, useContext } from "react";
 
 export function createSafeContext<TDdata>(init?: TDdata) {
   const context = createContext(init);
+
   const useSafeContext = () => {
     const ctx = useContext(context);
     if (ctx === undefined) {
@@ -9,5 +10,6 @@ export function createSafeContext<TDdata>(init?: TDdata) {
     }
     return ctx;
   };
+
   return [useSafeContext, context.Provider] as const;
 }
