+++ copilot-widget/lib/data/index.ts
@@ -1,0 +1,1 @@
+export * from "./chat";
