+++ copilot-widget/lib/hooks/useToggle.ts
@@ -1,6 +1,6 @@
 import { Dispatch, SetStateAction, useCallback, useState } from "react";
 
-export default function useToggle(
+export function useToggle(
   defaultValue?: boolean
 ): [boolean, () => void, Dispatch<SetStateAction<boolean>>] {
   const [value, setValue] = useState(!!defaultValue);
