+++ copilot-widget/lib/index.ts
@@ -1,22 +1,15 @@
-export { BotMessageWrapper } from "./components/Messages";
+export { BotMessageWrapper } from "./components/BotMessageWrapper";
 export { default as Root } from "./Root";
 export { CopilotWidget } from "./CopilotWidget";
-export type {
-  ComponentType,
-  ComponentProps,
-} from "./contexts/componentRegistery";
-export type { Options } from "./types/options";
-export {
-  useChatState,
-  useChatLoading,
-  useSendMessage,
-} from "./contexts/statefulMessageHandler";
-export { useConfigData } from "./contexts/ConfigData";
-export { useWidgetState } from "./contexts/WidgetState";
-export { useSocket } from "./contexts/SocketProvider";
-export { useLang } from "./contexts/LocalesProvider";
+
+export * from "./types";
+export * from "./hooks";
+export * from "./contexts";
+
 export type {
   UserMessageType,
   BotMessageType,
   HandoffPayloadType,
-} from "./contexts/messageHandler";
+  ComponentType,
+  ComponentProps,
+} from "./types";
