+++ copilot-widget/lib/locales/ar.locale.ts
@@ -1,4 +1,4 @@
-import { TranslatableMessages } from "./types";
+import { TranslatableMessages } from "@lib/types";
 
 export const arLocale: TranslatableMessages = {
   ok: "حسنا",
