+++ copilot-widget/lib/locales/en.locale.ts
@@ -1,4 +1,4 @@
-import { TranslatableMessages } from "./types";
+import { TranslatableMessages } from "@lib/types";
 
 export const enLocale: TranslatableMessages = {
   ok: "OK",
