+++ copilot-widget/lib/locales/index.ts
@@ -1,2 +1,1 @@
-export * from "./types";
 export * from "./helper";
