+++ copilot-widget/lib/types/index.ts
@@ -1,2 +1,6 @@
-export { type Message } from "./messageTypes";
-export { type Options } from "./options";
+export * from "./messageTypes";
+export * from "./helpers";
+export * from "./messages";
+export * from "./components";
+export * from "./options";
+export * from "./translations";
