+++ copilot-widget/lib/utils/utils.ts
@@ -14,12 +14,8 @@ function isEmpty(value: string | object | any[] | undefined): boolean {
   return false;
 }
 
-function getId(): string {
-  return Math.random().toString(36).substring(2, 15);
-}
-
 function getLast<T>(arr: T[]): T | undefined {
   return arr[arr.length - 1];
 }
 
-export { isEmpty, getId, getLast };
+export { isEmpty, getLast };
